+++ draft-js-emoji-plugin/src/components/EmojiSelect/Popover/index.js
@@ -17,7 +17,7 @@ export default class Popover extends Component {
     toneSelectOpenDelay: PropTypes.number.isRequired,
     isOpen: PropTypes.bool,
     useNativeArt: PropTypes.bool,
-    onSelect: PropTypes.func,
+    onSelect: PropTypes.func.isRequired,
   };
 
   static defaultProps = {
@@ -65,9 +65,7 @@ export default class Popover extends Component {
       emoji,
     );
     this.props.store.setEditorState(newEditorState);
-    if (this.props.onSelect) {
-      this.props.onSelect();
-    }
+    this.props.onSelect();
   };
 
   onEmojiMouseDown = (emojiEntry, toneSet) => {