+++ draft-js-emoji-plugin/CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
-- Added closePopupOnEmojiSelect option to configuration which when enabled closes emoji select popup after emoji has been clicked.
+- Emoji select popup now closes after emoji has been clicked.
 
 ## 2.0.6
 - Added an `id` attribute on the listbox options so the `aria-activedescendant` value refers to the focused option.